+++ src/image_ruiner.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+export default class ImageRuiner extends React.Component {
+    render() {
+        return (
+            <div>
+                loser
+            </div>
+        )
+    }
+}
