+++ src/index.js
@@ -1,47 +1,29 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
+import {
+    createBrowserRouter,
+    RouterProvider,
+} from "react-router-dom"
 import './index.css';
 
+import ErrorPage from "./error_page"
+import ImageRuiner from "./image_ruiner"
+import SentenceBuilder from "./sentece_builder"
 
-class SentenceBuilder extends React.Component {
-    constructor() {
-        super()
-        this.state = {
-            adjective: null,
-            noun: null
-        }
-    }
 
-    componentDidMount() {
-        fetch('https://gist.githubusercontent.com/hugsy/8910dc78d208e40de42deb29e62df913/raw/eec99c5597a73f6a9240cab26965a8609fa0f6ea/english-nouns.txt')
-            .then((response) => response.text())
-            .then((data) => {
-                let split_data = data.split("\n")
-                console.log(split_data)
-                this.setState({ noun: split_data[Math.floor(Math.random() * split_data.length)] })
-            })
+const router = createBrowserRouter([{
+    path: '/',
+    element: <SentenceBuilder />,
+    errorElement: <ErrorPage />
+}, {
+    path: '/test',
+    element: <ImageRuiner />
+},])
 
-        fetch('https://gist.githubusercontent.com/hugsy/8910dc78d208e40de42deb29e62df913/raw/eec99c5597a73f6a9240cab26965a8609fa0f6ea/english-adjectives.txt')
-            .then((response) => response.text())
-            .then((data) => {
-                let split_data = data.split("\n")
-                this.setState({ adjective: split_data[Math.floor(Math.random() * split_data.length)] })
-            })
-    }
-
-    render() {
-        return (
-            <div>
-                <a href="https://github.com/ramonmeza">
-                    <img src="images/profile_pic.jpg" alt="It's a-me" />
-                </a>
-                <p>
-                    Ramon is {(this.state.adjective && this.state.adjective.match('^[aieouAIEOU].*')) ? 'an' : 'a'} <span class="adjective">{this.state.adjective}</span> <span class="noun">{this.state.noun}</span>.
-                </p >
-            </div>
-        )
-    }
-}
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
-root.render(<SentenceBuilder />);
+root.render(
+    <React.StrictMode>
+        <RouterProvider router={router} />
+    </React.StrictMode>
+)
